+++ packages/material-ui-system/src/style.test.js
@@ -63,6 +63,37 @@ describe('style', () => {
     // });
   });
 
+  const boxShadow = style({
+    prop: 'boxShadow',
+    themeKey: 'shadows',
+  });
+
+  it('should support array theme value', () => {
+    const output = boxShadow({
+      theme: {
+        shadows: ['none', '0px 1px 3px 0px rgba(0, 0, 0, 0.2)'],
+      },
+      boxShadow: 1,
+    });
+
+    assert.deepEqual(output, {
+      boxShadow: '0px 1px 3px 0px rgba(0, 0, 0, 0.2)',
+    });
+  });
+
+  it('should fallback to value if theme value is an array and index missing', () => {
+    const output = boxShadow({
+      theme: {
+        shadows: ['none'],
+      },
+      boxShadow: '0px 1px 3px 0px rgba(0, 0, 0, 0.2)',
+    });
+
+    assert.deepEqual(output, {
+      boxShadow: '0px 1px 3px 0px rgba(0, 0, 0, 0.2)',
+    });
+  });
+
   const border = style({
     prop: 'border',
     themeKey: 'borders',
